Separate Catch, Realtime, Supervisor, and BlockFetcher states
BlockFetcher.Catchup and BlockFetcher.Realtime get their own structs,
which use BlockFetcher struct.  This better emphasizes that Catchup and
Realtime are both block fetchers.  The BlockFetcher.Catchup and
BlockFetcher.Realtime structs move into Supervisor's struct.
