+++ apps/indexer/lib/indexer/block_fetcher/realtime.ex
@@ -5,29 +5,46 @@ defmodule Indexer.BlockFetcher.Realtime do
 
   require Logger
 
-  import Indexer.BlockFetcher, only: [stream_import: 4]
+  import Indexer.BlockFetcher, only: [stream_import: 1]
 
   alias Indexer.{BlockFetcher, Sequence}
+
+  @enforce_keys ~w(block_fetcher interval)a
+  defstruct block_fetcher: nil,
+            interval: nil,
+            task_by_ref: %{}
+
+  def new(%{block_fetcher: %BlockFetcher{} = common_block_fetcher, block_interval: block_interval}) do
+    block_fetcher = %BlockFetcher{
+      block_fetcher | blocks_concurrency: 1, broadcast: true}
+
+    interval = div(block_interval, 2)
+
+    %__MODULE__{block_fetcher: block_fetcher, interval: interval}
+  end
 
   @doc """
   Starts `task/1` and puts it in `t:Indexer.BlockFetcher.t/0` `realtime_task_by_ref`.
   """
-  def put(%BlockFetcher{} = state) do
-    %Task{ref: ref} = realtime_task = Task.Supervisor.async_nolink(Indexer.TaskSupervisor, __MODULE__, :task, [state])
+  def put(%BlockFetcher.Supervisor{realtime: %__MODULE__{} = state} = supervisor_state) do
+    %Task{ref: ref} = task = Task.Supervisor.async_nolink(Indexer.TaskSupervisor, __MODULE__, :task, [state])
 
-    put_in(state.realtime_task_by_ref[ref], realtime_task)
+    put_in(supervisor_state.realtime.task_by_ref[ref], task)
   end
 
-  def task(%BlockFetcher{json_rpc_named_arguments: json_rpc_named_arguments} = state) do
+  def task(%__MODULE__{block_fetcher: %BlockFetcher{json_rpc_named_arguments: json_rpc_named_arguments} = block_fetcher}) do
     {:ok, latest_block_number} = EthereumJSONRPC.fetch_block_number_by_tag("latest", json_rpc_named_arguments)
-    {:ok, seq} = Sequence.start_link(first: latest_block_number, step: 2)
-    stream_import(state, seq, :realtime_index, max_concurrency: 1)
+    {:ok, sequence} = Sequence.start_link(first: latest_block_number, step: 2)
+    stream_import(%BlockFetcher{block_fetcher | sequence: sequence})
   end
 
-  def handle_success({ref, :ok = result}, %BlockFetcher{realtime_task_by_ref: realtime_task_by_ref} = state) do
-    {realtime_task, running_realtime_task_by_ref} = Map.pop(realtime_task_by_ref, ref)
+  def handle_success(
+        {ref, :ok = result},
+        %BlockFetcher.Supervisor{realtime: %__MODULE__{task_by_ref: task_by_ref}} = supervisor_state
+      ) do
+    {task, running_task_by_ref} = Map.pop(task_by_ref, ref)
 
-    case realtime_task do
+    case task do
       nil ->
         Logger.error(fn ->
           "Unknown ref (#{inspect(ref)}) that is neither the catchup index" <>
@@ -40,16 +57,16 @@ defmodule Indexer.BlockFetcher.Realtime do
 
     Process.demonitor(ref, [:flush])
 
-    %BlockFetcher{state | realtime_task_by_ref: running_realtime_task_by_ref}
+    put_in(supervisor_state.realtime.task_by_ref, running_task_by_ref)
   end
 
   def handle_failure(
         {:DOWN, ref, :process, pid, reason},
-        %BlockFetcher{realtime_task_by_ref: realtime_task_by_ref} = state
+        %BlockFetcher.Supervisor{realtime: %__MODULE__{task_by_ref: task_by_ref}} = supervisor_state
       ) do
-    {realtime_task, running_realtime_task_by_ref} = Map.pop(realtime_task_by_ref, ref)
+    {task, running_task_by_ref} = Map.pop(task_by_ref, ref)
 
-    case realtime_task do
+    case task do
       nil ->
         Logger.error(fn ->
           "Unknown ref (#{inspect(ref)}) that is neither the catchup index" <>
@@ -64,6 +81,6 @@ defmodule Indexer.BlockFetcher.Realtime do
         end)
     end
 
-    %BlockFetcher{state | realtime_task_by_ref: running_realtime_task_by_ref}
+    put_in(supervisor_state.realtime.task_by_ref, running_task_by_ref)
   end
 end
